+ Setup Direct3D 11 renderer

+++ hw3d/src/WinMain.cpp
@@ -9,5 +9,5 @@ int WINAPI WinMain(
     int cmdshow)
 {
     Application app{};
-    return app.Start();
+    return (int)app.Start();
 }
